refactor: organize files for substrate
